perf(tests): extract searchInTable as pure function and move tests to tableValues
